change rolling window history to use Vec instead of VecDeque

+++ ff_standard_lib/src/consolidators/candlesticks.rs
@@ -279,7 +279,7 @@ impl CandleStickConsolidator {
         self.history.clone()
     }
 
-    pub(crate) fn index(&self, index: u64) -> Option<BaseDataEnum> {
+    pub(crate) fn index(&self, index: usize) -> Option<BaseDataEnum> {
         match self.history.get(index) {
             Some(data) => Some(data.clone()),
             None => None,
