+++ ff_standard_lib/src/consolidators/heikinashi.rs
@@ -340,7 +340,7 @@ impl HeikinAshiConsolidator {
         self.history.clone()
     }
 
-    pub(crate) fn index(&self, index: u64) -> Option<BaseDataEnum> {
+    pub(crate) fn index(&self, index: usize) -> Option<BaseDataEnum> {
         match self.history.get(index) {
             Some(data) => Some(data.clone()),
             None => None,
